requirements.txt
certifi==2022.9.24
cryptography==38.0.4
Flask==2.2.2
Flask_Cors==3.0.10
Flask_RESTful==0.3.9
pydantic==1.10.4
PyJWT==2.6.0
pymongo==4.3.3
python-dotenv==0.21.0
Werkzeug==2.2.2
